refactor(cell/cell_methods): split inventory/tests.rs by behaviour (#244)
inventory/tests.rs was 770 lines (over the 700-line cap). Split by
the inventory message family each test exercises:

  tests/mod.rs               - shared make_test_space_mgr + register_test_fire_ability helpers
  tests/slot_swap.rs         - slot-swap ammo preservation + reload cancellation (2 tests)
  tests/ammo_change.rs       - REQUEST_AMMO_CHANGE happy-path + validation (3 tests)
  tests/active_slot_change.rs- REQUEST_ACTIVE_SLOT_CHANGE bounds + wire/server translation (2 tests)
  tests/move_item.rs         - MOVE_ITEM wire/server slot translation (1 test)

All 8 tests pass; no behavior change.
